Master

+++ src/main/java/athenarc/imsi/sdl/web/rest/DatasetsResource.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 import org.bson.Document;
@@ -24,6 +25,11 @@ import org.springframework.web.multipart.MultipartFile;
 
 import athenarc.imsi.sdl.service.DatasetsService;
 import athenarc.imsi.sdl.service.util.FileUtil;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import springfox.documentation.annotations.ApiIgnore;
+
 /**
  * RankingResource controller
  */
@@ -32,42 +38,65 @@ import athenarc.imsi.sdl.service.util.FileUtil;
 public class DatasetsResource {
 
     private final Logger log = LoggerFactory.getLogger(DatasetsResource.class);
-    
+
     @Autowired
     private final DatasetsService datasetsService = new DatasetsService();
-    
+
     /**
-    * POST submit
-    */
+     * POST submit
+     */
+    @ApiIgnore
     @PostMapping("/upload")
     public Document upload(@RequestParam("file") MultipartFile file) {
-            String filename = file.getOriginalFilename();
+        String filename = file.getOriginalFilename();
 
-            if (file.isEmpty()) {
-                throw new RuntimeException("Error uploading " + filename);
+        if (file.isEmpty()) {
+            throw new RuntimeException("Error uploading " + filename);
+        }
+
+        try {
+
+            String zipFile = datasetsService.upload(filename, file.getBytes());
+
+            List<String> initialDatasets = FileUtil.getLocalDatasets();
+
+            if (FileUtil.unzip(zipFile) != 0) {
+                throw new RuntimeException("Error unzipping " + zipFile);
             }
 
-            try {
-                String zipFile = datasetsService.upload(filename, file.getBytes());
-                if (FileUtil.unzip(zipFile) != 0){
-                    throw new RuntimeException("Error unzipping " + zipFile);
-                }
+            List<String> currentDatasets = FileUtil.getLocalDatasets();
 
-                if (!FileUtil.remove(zipFile)) {
-                    throw new RuntimeException("Error removing zip  " + zipFile);
-                }
+            // find the folder that was just created - though unzip
+            List<String> diff = new ArrayList(currentDatasets);
+            diff.removeAll(initialDatasets);
+            String newDataset = diff.get(0);
 
-            } catch (IOException | InterruptedException e) {
-                throw new RuntimeException("Error uploading " + filename);
+            // copy that folder to hdfs
+            if (FileUtil.copyToHdfs(newDataset) != 0) {
+                throw new RuntimeException("Error copying to HDFS " + zipFile);
             }
 
-            return new Document("status", "success");
+            if (!FileUtil.remove(zipFile)) {
+                throw new RuntimeException("Error removing zip  " + zipFile);
+            }
+
+        } catch (IOException | InterruptedException e) {
+            throw new RuntimeException("Error uploading " + filename);
+        }
+
+        return new Document("status", "success");
     }
 
     /**
-    * GET status
-    */
-    @GetMapping("/schemas")
+     * GET status
+     */
+    @ApiOperation(value = "Used to retrieve the names and schemas of the available datasets")
+    @ApiResponses(value =
+        {
+            @ApiResponse(code = 200, message = "Responds with a dictionary, mapping each available dataset to its elements")
+        }
+    )
+    @GetMapping(value = "/schemas", produces = "application/json;charset=UTF-8")
     public Document getSchemas() {
         Document response;
         try {
@@ -80,21 +109,39 @@ public class DatasetsResource {
     }
 
     /**
-    * GET status
-    */
+     * GET status
+     */
+    @ApiOperation(value = "Used for searching a dataset and retrieving field values from a specific dataset entity, that contain a given literal")
+    @ApiResponses(value =
+        {
+            @ApiResponse(code = 200, message = "Responds with a list of values")
+        }
+    )
     @GetMapping("/autocomplete")
-    public List<Document> autocomplete(@RequestParam(value = "folder") String folder, 
-        @RequestParam(value = "entity") String entity, 
-        @RequestParam(value = "field") String field,
-        @RequestParam(value = "term") String term) {
+    public List<Document> autocomplete(@RequestParam(value = "term") String term) {
 
+        // give terms to service so that it can check whether the all of the terms exist
+        // in case of an error during the service execution throw a runtime error
+        // in case of a successful service execution, return the service result
         try {
-            return datasetsService.autocomplete(folder, entity.substring(0, 1), field, term.toLowerCase());
+            return datasetsService.autocomplete(term.toLowerCase());
         } catch (IOException e) {
             throw new RuntimeException("Error reading schema for datasets");
         }
     }
 
+    /**
+    * GET status
+    */
+    @GetMapping("/get")
+    public Document get(@RequestParam(value = "term") String term) {
+
+        try {
+            return datasetsService.get(term.toLowerCase());
+        } catch (IOException e) {
+            throw new RuntimeException("Error reading schema for datasets");
+        }
+    }
     /**
     * GET download result
     */
@@ -102,20 +149,20 @@ public class DatasetsResource {
     public ResponseEntity<Resource> download(String analysisType, String id) {
 
         try {
-        String downloadFile = FileUtil.getOutputFile(id, analysisType);
-        File fd = new File(downloadFile);
-        InputStreamResource resource = new InputStreamResource(new FileInputStream(fd));
+            String downloadFile = FileUtil.getOutputFile(id, analysisType);
+            File fd = new File(downloadFile);
+            InputStreamResource resource = new InputStreamResource(new FileInputStream(fd));
 
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Cache-Control", "no-cache, no-store, must-revalidate");
-        headers.add("Pragma", "no-cache");
-        headers.add("Expires", "0");
+            HttpHeaders headers = new HttpHeaders();
+            headers.add("Cache-Control", "no-cache, no-store, must-revalidate");
+            headers.add("Pragma", "no-cache");
+            headers.add("Expires", "0");
 
-        return ResponseEntity.ok()
-            .headers(headers)
-            .contentLength(fd.length())
-            .contentType(MediaType.parseMediaType("application/octet-stream"))
-            .body(resource);
+            return ResponseEntity.ok()
+                .headers(headers)
+                .contentLength(fd.length())
+                .contentType(MediaType.parseMediaType("application/octet-stream"))
+                .body(resource);
 
         } catch (FileNotFoundException e) {
             throw new RuntimeException("Error downloading result file");
